Follow eslint's advice & don't use a namespace
